Draft: move functions to draftgeoutils.arcs
